Move components to folder
