+++ src/views/GovernanceReviewTeam/Actions/ActionForm.tsx
@@ -15,7 +15,7 @@ import { SystemIntakeContactProps } from 'types/systemIntake';
 import Breadcrumbs from 'views/TechnicalAssistance/Breadcrumbs';
 import Pager from 'views/TechnicalAssistance/RequestForm/Pager';
 
-import EmailRecipientsFields from './EmailRecipientsFields';
+import EmailRecipientsFields from './components/EmailRecipientsFields';
 
 // TODO: update fields to match schema when backend work is completed
 export interface SystemIntakeActionFields {
